refactor: Use env veriables for rgbpp spore examples
