--- examples/rgbpp/local/3-btc-jump-ckb.ts
@@ -1,117 +1,0 @@
-import { serializeScript } from '@nervosnetwork/ckb-sdk-utils';
-import {
-  Collector,
-  appendCkbTxWitnesses,
-  genBtcJumpCkbVirtualTx,
-  sendCkbTx,
-  getXudtTypeScript,
-  updateCkbTxWithRealBtcTxId,
-  buildRgbppLockArgs,
-} from '@rgbpp-sdk/ckb';
-import { sendRgbppUtxos, DataSource, NetworkType, bitcoin, ECPair, transactionToHex } from '@rgbpp-sdk/btc';
-import { BtcAssetsApi, BtcAssetsApiError } from '@rgbpp-sdk/service';
-
-// BTC SECP256K1 private key
-const BTC_TEST_PRIVATE_KEY = '0000000000000000000000000000000000000000000000000000000000000001';
-// API docs: https://btc-assets-api.testnet.mibao.pro/docs
-const BTC_ASSETS_API_URL = 'https://btc-assets-api.testnet.mibao.pro';
-// https://btc-assets-api.testnet.mibao.pro/docs/static/index.html#/Token/post_token_generate
-const BTC_ASSETS_TOKEN = '';
-
-const BTC_ASSETS_ORIGIN = 'https://btc-test.app';
-
-interface Params {
-  rgbppLockArgsList: string[];
-  toCkbAddress: string;
-  transferAmount: bigint;
-}
-const jumpFromBtcToCkb = async ({ rgbppLockArgsList, toCkbAddress, transferAmount }: Params) => {
-  const collector = new Collector({
-    ckbNodeUrl: 'https://testnet.ckb.dev/rpc',
-    ckbIndexerUrl: 'https://testnet.ckb.dev/indexer',
-  });
-  const isMainnet = false;
-
-  const network = isMainnet ? bitcoin.networks.bitcoin : bitcoin.networks.testnet;
-  const keyPair = ECPair.fromPrivateKey(Buffer.from(BTC_TEST_PRIVATE_KEY, 'hex'), { network });
-  const { address: btcAddress } = bitcoin.payments.p2wpkh({
-    pubkey: keyPair.publicKey,
-    network,
-  });
-  console.log('btc address: ', btcAddress);
-
-  const networkType = isMainnet ? NetworkType.MAINNET : NetworkType.TESTNET;
-  const service = BtcAssetsApi.fromToken(BTC_ASSETS_API_URL, BTC_ASSETS_TOKEN, BTC_ASSETS_ORIGIN);
-  const source = new DataSource(service, networkType);
-
-  // Warning: Please replace with your real xUDT type script here
-  const xudtType: CKBComponents.Script = {
-    ...getXudtTypeScript(isMainnet),
-    args: '0x1ba116c119d1cfd98a53e9d1a615cf2af2bb87d95515c9d217d367054cfc696b',
-  };
-
-  const ckbVirtualTxResult = await genBtcJumpCkbVirtualTx({
-    collector,
-    rgbppLockArgsList,
-    xudtTypeBytes: serializeScript(xudtType),
-    transferAmount,
-    toCkbAddress,
-    isMainnet,
-  });
-
-  const { commitment, ckbRawTx } = ckbVirtualTxResult;
-
-  // Send BTC tx
-  const psbt = await sendRgbppUtxos({
-    ckbVirtualTx: ckbRawTx,
-    commitment,
-    tos: [btcAddress!],
-    ckbCollector: collector,
-    from: btcAddress!,
-    source,
-  });
-  psbt.signAllInputs(keyPair);
-  psbt.finalizeAllInputs();
-
-  const btcTx = psbt.extractTransaction();
-  // Remove the witness from BTC tx for RGBPP unlock
-  const btcTxBytes = transactionToHex(btcTx, false);
-  const { txid: btcTxId } = await service.sendBtcTransaction(btcTx.toHex());
-
-  console.log('BTC Tx bytes: ', btcTxBytes);
-  console.log('BTC TxId: ', btcTxId);
-  console.log('ckbRawTx', JSON.stringify(ckbRawTx));
-
-  // Wait for BTC tx and proof to be ready, and then send isomorphic CKB transactions
-  const interval = setInterval(async () => {
-    try {
-      console.log('Waiting for BTC tx and proof to be ready');
-      const rgbppApiSpvProof = await service.getRgbppSpvProof(btcTxId, 0);
-      clearInterval(interval);
-      // Update CKB transaction with the real BTC txId
-      const newCkbRawTx = updateCkbTxWithRealBtcTxId({ ckbRawTx, btcTxId, isMainnet });
-      const ckbTx = await appendCkbTxWitnesses({
-        ckbRawTx: newCkbRawTx,
-        btcTxBytes,
-        rgbppApiSpvProof,
-      });
-
-      const txHash = await sendCkbTx({ collector, signedTx: ckbTx });
-      console.info(`RGB++ Asset has been leaped from BTC to CKB and the CKB tx hash is ${txHash}`);
-    } catch (error) {
-      if (!(error instanceof BtcAssetsApiError)) {
-        console.error(error);
-      }
-    }
-  }, 30 * 1000);
-};
-
-// rgbppLockArgs: outIndexU32 + btcTxId
-jumpFromBtcToCkb({
-  // If the `3-btc-transfer.ts` has been executed, the BTC txId should be the new generated BTC txId by the `3-btc-transfer.ts`
-  // Otherwise the BTC txId should be same as the the BTC txId of the `2-ckb-jump-btc.ts`
-  rgbppLockArgsList: [buildRgbppLockArgs(1, '24e622419156dd3a277a90bcbb40c7117462a18d5329dd1ada320ca8bdfba715')],
-  toCkbAddress: 'ckt1qrfrwcdnvssswdwpn3s9v8fp87emat306ctjwsm3nmlkjg8qyza2cqgqq9kxr7vy7yknezj0vj0xptx6thk6pwyr0sxamv6q',
-  // To simplify, keep the transferAmount the same as 2-ckb-jump-btc
-  transferAmount: BigInt(800_0000_0000),
-});
