--- examples/rgbpp/local/launch/0-rgbpp-token-info.ts
@@ -1,7 +1,0 @@
-import { RgbppTokenInfo } from '@rgbpp-sdk/ckb';
-
-export const RGBPP_TOKEN_INFO: RgbppTokenInfo = {
-  decimal: 8,
-  name: 'RGBPP Test Token',
-  symbol: 'RTT',
-};
