+++ examples/rgbpp/spore/4-transfer-spore.ts
@@ -1,57 +1,21 @@
-import {
-  Collector,
-  buildRgbppLockArgs,
-  appendCkbTxWitnesses,
-  updateCkbTxWithRealBtcTxId,
-  sendCkbTx,
-  getSporeTypeScript,
-  Hex,
-  generateSporeTransferCoBuild,
-  genTransferSporeCkbVirtualTx,
-} from '@rgbpp-sdk/ckb';
-import { DataSource, ECPair, bitcoin, NetworkType, sendRgbppUtxos, transactionToHex } from '@rgbpp-sdk/btc';
-import { BtcAssetsApi, BtcAssetsApiError } from '@rgbpp-sdk/service';
+import { buildRgbppLockArgs, getSporeTypeScript, Hex, genTransferSporeCkbVirtualTx } from '@rgbpp-sdk/ckb';
+import { sendRgbppUtxos } from '@rgbpp-sdk/btc';
 import { serializeScript } from '@nervosnetwork/ckb-sdk-utils';
-
-// BTC SECP256K1 private key
-const BTC_TEST_PRIVATE_KEY = '0000000000000000000000000000000000000000000000000000000000000001';
-// API docs: https://btc-assets-api.testnet.mibao.pro/docs
-const BTC_ASSETS_API_URL = 'https://btc-assets-api.testnet.mibao.pro';
-// https://btc-assets-api.testnet.mibao.pro/docs/static/index.html#/Token/post_token_generate
-const BTC_ASSETS_TOKEN = '';
-
-const BTC_ASSETS_ORIGIN = 'https://btc-test.app';
+import { isMainnet, collector, btcAddress, btcDataSource, btcKeyPair, btcService } from 'examples-core';
 
 const transferSpore = async ({
   sporeRgbppLockArgs,
   toBtcAddress,
+  sporeTypeArgs,
 }: {
   sporeRgbppLockArgs: Hex;
   toBtcAddress: string;
+  sporeTypeArgs: Hex;
 }) => {
-  const collector = new Collector({
-    ckbNodeUrl: 'https://testnet.ckb.dev/rpc',
-    ckbIndexerUrl: 'https://testnet.ckb.dev/indexer',
-  });
-  const isMainnet = false;
-
-  const network = isMainnet ? bitcoin.networks.bitcoin : bitcoin.networks.testnet;
-  const keyPair = ECPair.fromPrivateKey(Buffer.from(BTC_TEST_PRIVATE_KEY, 'hex'), { network });
-  const { address: btcAddress } = bitcoin.payments.p2wpkh({
-    pubkey: keyPair.publicKey,
-    network,
-  });
-
-  console.log('btc address: ', btcAddress);
-
-  const networkType = isMainnet ? NetworkType.MAINNET : NetworkType.TESTNET;
-  const service = BtcAssetsApi.fromToken(BTC_ASSETS_API_URL, BTC_ASSETS_TOKEN, BTC_ASSETS_ORIGIN);
-  const source = new DataSource(service, networkType);
-
-  // The spore type script is from 3-create-spore.ts, you can find it from the ckb tx spore output cells
+  // The spore type script is from 3-create-spore.ts, you can find it from the ckb tx spore output cell
   const sporeTypeBytes = serializeScript({
     ...getSporeTypeScript(isMainnet),
-    args: '0x205fe15af04e59d3ff1ff8e0b0a1e3bc201af406a38964760c24848ed6029b6b',
+    args: sporeTypeArgs,
   });
 
   const ckbVirtualTxResult = await genTransferSporeCkbVirtualTx({
@@ -59,10 +23,9 @@ const transferSpore = async ({
     sporeRgbppLockArgs,
     sporeTypeBytes,
     isMainnet,
-    ckbFeeRate: BigInt(5000),
   });
 
-  const { commitment, ckbRawTx, sporeCell } = ckbVirtualTxResult;
+  const { commitment, ckbRawTx } = ckbVirtualTxResult;
 
   // console.log(JSON.stringify(ckbRawTx))
 
@@ -73,51 +36,42 @@ const transferSpore = async ({
     tos: [toBtcAddress],
     ckbCollector: collector,
     from: btcAddress!,
-    source,
+    source: btcDataSource,
     feeRate: 30,
   });
-  psbt.signAllInputs(keyPair);
+  psbt.signAllInputs(btcKeyPair);
   psbt.finalizeAllInputs();
 
   const btcTx = psbt.extractTransaction();
-  const btcTxBytes = transactionToHex(btcTx, false);
-  const { txid: btcTxId } = await service.sendBtcTransaction(btcTx.toHex());
+  const { txid: btcTxId } = await btcService.sendBtcTransaction(btcTx.toHex());
 
   console.log('BTC TxId: ', btcTxId);
 
-  const interval = setInterval(async () => {
-    try {
-      console.log('Waiting for BTC tx and proof to be ready');
-      const rgbppApiSpvProof = await service.getRgbppSpvProof(btcTxId, 0);
-      clearInterval(interval);
-      // Update CKB transaction with the real BTC txId
-      const newCkbRawTx = updateCkbTxWithRealBtcTxId({ ckbRawTx, btcTxId, isMainnet });
-
-      const ckbTx = await appendCkbTxWitnesses({
-        ckbRawTx: newCkbRawTx,
-        btcTxBytes,
-        rgbppApiSpvProof,
-      });
-
-      // Replace cobuild witness with the final rgbpp lock script
-      ckbTx.witnesses[ckbTx.witnesses.length - 1] = generateSporeTransferCoBuild([sporeCell], ckbTx.outputs);
-
-      // console.log('ckbTx: ', JSON.stringify(ckbTx));
-
-      const txHash = await sendCkbTx({ collector, signedTx: ckbTx });
-      console.info(`RGB++ Spore has been transferred and tx hash is ${txHash}`);
-    } catch (error) {
-      if (!(error instanceof BtcAssetsApiError)) {
-        console.error(error);
+  try {
+    await btcService.sendRgbppCkbTransaction({ btc_txid: btcTxId, ckb_virtual_result: ckbVirtualTxResult });
+    const interval = setInterval(async () => {
+      const { state, failedReason } = await btcService.getRgbppTransactionState(btcTxId);
+      console.log('state', state);
+      if (state === 'completed' || state === 'failed') {
+        clearInterval(interval);
+        if (state === 'completed') {
+          const { txhash: txHash } = await btcService.getRgbppTransactionHash(btcTxId);
+          console.info(`Rgbpp spore has been transferred on BTC and the related CKB tx hash is ${txHash}`);
+        } else {
+          console.warn(`Rgbpp CKB transaction failed and the reason is ${failedReason} `);
+        }
       }
-    }
-  }, 30 * 1000);
+    }, 30 * 1000);
+  } catch (error) {
+    console.error(error);
+  }
 };
 
 // Use your real BTC UTXO information on the BTC Testnet
 // rgbppLockArgs: outIndexU32 + btcTxId
 transferSpore({
   // The spore rgbpp lock args is from 3-create-spore.ts
-  sporeRgbppLockArgs: buildRgbppLockArgs(1, 'f203c8c13eacdbd126f85d286a963c85f233f8145363b1d997c4d552afb990e1'),
+  sporeRgbppLockArgs: buildRgbppLockArgs(2, 'd5868dbde4be5e49876b496449df10150c356843afb6f94b08f8d81f394bb350'),
   toBtcAddress: 'tb1qhp9fh9qsfeyh0yhewgu27ndqhs5qlrqwau28m7',
+  sporeTypeArgs: '0x42898ea77062256f46e8f1b861d526ae47810ecc51ab50477945d5fa90452706',
 });
