+++ examples/rgbpp/spore/local/4-transfer-spore.ts
@@ -3,32 +3,48 @@ import {
   appendCkbTxWitnesses,
   updateCkbTxWithRealBtcTxId,
   sendCkbTx,
-  genCreateClusterCkbVirtualTx,
-  generateClusterCreateCoBuild,
+  getSporeTypeScript,
+  Hex,
+  generateSporeTransferCoBuild,
+  genTransferSporeCkbVirtualTx,
 } from '@rgbpp-sdk/ckb';
 import { sendRgbppUtxos, transactionToHex } from '@rgbpp-sdk/btc';
 import { BtcAssetsApiError } from '@rgbpp-sdk/service';
+import { serializeScript } from '@nervosnetwork/ckb-sdk-utils';
 import { isMainnet, collector, btcAddress, btcDataSource, btcKeyPair, btcService } from 'examples-core';
-import { CLUSTER_DATA } from './0-cluster-info';
 
-const createCluster = async ({ ownerRgbppLockArgs }: { ownerRgbppLockArgs: string }) => {
-  const ckbVirtualTxResult = await genCreateClusterCkbVirtualTx({
+const transferSpore = async ({
+  sporeRgbppLockArgs,
+  toBtcAddress,
+  sporeTypeArgs,
+}: {
+  sporeRgbppLockArgs: Hex;
+  toBtcAddress: string;
+  sporeTypeArgs: string;
+}) => {
+  // The spore type script is from 3-create-spore.ts, you can find it from the ckb tx spore output cells
+  const sporeTypeBytes = serializeScript({
+    ...getSporeTypeScript(isMainnet),
+    args: sporeTypeArgs,
+  });
+
+  const ckbVirtualTxResult = await genTransferSporeCkbVirtualTx({
     collector,
-    rgbppLockArgs: ownerRgbppLockArgs,
-    clusterData: CLUSTER_DATA,
+    sporeRgbppLockArgs,
+    sporeTypeBytes,
     isMainnet,
     ckbFeeRate: BigInt(5000),
   });
 
-  const { commitment, ckbRawTx, clusterId } = ckbVirtualTxResult;
+  const { commitment, ckbRawTx, sporeCell } = ckbVirtualTxResult;
 
-  console.log('clusterId: ', clusterId);
+  // console.log(JSON.stringify(ckbRawTx))
 
   // Send BTC tx
   const psbt = await sendRgbppUtxos({
     ckbVirtualTx: ckbRawTx,
     commitment,
-    tos: [btcAddress!],
+    tos: [toBtcAddress],
     ckbCollector: collector,
     from: btcAddress!,
     source: btcDataSource,
@@ -50,21 +66,20 @@ const createCluster = async ({ ownerRgbppLockArgs }: { ownerRgbppLockArgs: strin
       clearInterval(interval);
       // Update CKB transaction with the real BTC txId
       const newCkbRawTx = updateCkbTxWithRealBtcTxId({ ckbRawTx, btcTxId, isMainnet });
+
       const ckbTx = await appendCkbTxWitnesses({
         ckbRawTx: newCkbRawTx,
         btcTxBytes,
         rgbppApiSpvProof,
       });
-      // Replace cobuild witness with the final rgbpp lock script
-      ckbTx.witnesses[ckbTx.witnesses.length - 1] = generateClusterCreateCoBuild(
-        ckbTx.outputs[0],
-        ckbTx.outputsData[0],
-      );
 
-      console.log(JSON.stringify(ckbTx));
+      // Replace cobuild witness with the final rgbpp lock script
+      ckbTx.witnesses[ckbTx.witnesses.length - 1] = generateSporeTransferCoBuild([sporeCell], ckbTx.outputs);
+
+      // console.log('ckbTx: ', JSON.stringify(ckbTx));
 
       const txHash = await sendCkbTx({ collector, signedTx: ckbTx });
-      console.info(`RGB++ Cluster has been created and tx hash is ${txHash}`);
+      console.info(`RGB++ Spore has been transferred and tx hash is ${txHash}`);
     } catch (error) {
       if (!(error instanceof BtcAssetsApiError)) {
         console.error(error);
@@ -75,6 +90,9 @@ const createCluster = async ({ ownerRgbppLockArgs }: { ownerRgbppLockArgs: strin
 
 // Use your real BTC UTXO information on the BTC Testnet
 // rgbppLockArgs: outIndexU32 + btcTxId
-createCluster({
-  ownerRgbppLockArgs: buildRgbppLockArgs(3, 'aee4e8e3aa95e9e9ab1f0520714031d92d3263262099dcc7f7d64e62fa2fcb44'),
+transferSpore({
+  // The spore rgbpp lock args is from 3-create-spore.ts
+  sporeRgbppLockArgs: buildRgbppLockArgs(1, 'f203c8c13eacdbd126f85d286a963c85f233f8145363b1d997c4d552afb990e1'),
+  toBtcAddress: 'tb1qhp9fh9qsfeyh0yhewgu27ndqhs5qlrqwau28m7',
+  sporeTypeArgs: '0x42898ea77062256f46e8f1b861d526ae47810ecc51ab50477945d5fa90452706',
 });
