--- examples/rgbpp/spore/queue/4-transfer-spore.ts
@@ -1,102 +1,0 @@
-import { Collector, buildRgbppLockArgs, getSporeTypeScript, Hex, genTransferSporeCkbVirtualTx } from '@rgbpp-sdk/ckb';
-import { DataSource, ECPair, bitcoin, NetworkType, sendRgbppUtxos } from '@rgbpp-sdk/btc';
-import { BtcAssetsApi } from '@rgbpp-sdk/service';
-import { serializeScript } from '@nervosnetwork/ckb-sdk-utils';
-
-// BTC SECP256K1 private key
-const BTC_TEST_PRIVATE_KEY = '0000000000000000000000000000000000000000000000000000000000000001';
-// API docs: https://btc-assets-api.testnet.mibao.pro/docs
-const BTC_ASSETS_API_URL = 'https://btc-assets-api.testnet.mibao.pro';
-// https://btc-assets-api.testnet.mibao.pro/docs/static/index.html#/Token/post_token_generate
-const BTC_ASSETS_TOKEN = '';
-
-const BTC_ASSETS_ORIGIN = 'https://btc-test.app';
-
-const transferSpore = async ({
-  sporeRgbppLockArgs,
-  toBtcAddress,
-}: {
-  sporeRgbppLockArgs: Hex;
-  toBtcAddress: string;
-}) => {
-  const collector = new Collector({
-    ckbNodeUrl: 'https://testnet.ckb.dev/rpc',
-    ckbIndexerUrl: 'https://testnet.ckb.dev/indexer',
-  });
-  const isMainnet = false;
-
-  const network = isMainnet ? bitcoin.networks.bitcoin : bitcoin.networks.testnet;
-  const keyPair = ECPair.fromPrivateKey(Buffer.from(BTC_TEST_PRIVATE_KEY, 'hex'), { network });
-  const { address: btcAddress } = bitcoin.payments.p2wpkh({
-    pubkey: keyPair.publicKey,
-    network,
-  });
-
-  console.log('btc address: ', btcAddress);
-
-  const networkType = isMainnet ? NetworkType.MAINNET : NetworkType.TESTNET;
-  const service = BtcAssetsApi.fromToken(BTC_ASSETS_API_URL, BTC_ASSETS_TOKEN, BTC_ASSETS_ORIGIN);
-  const source = new DataSource(service, networkType);
-
-  // The spore type script is from 3-create-spore.ts, you can find it from the ckb tx spore output cell
-  const sporeTypeBytes = serializeScript({
-    ...getSporeTypeScript(isMainnet),
-    args: '0x205fe15af04e59d3ff1ff8e0b0a1e3bc201af406a38964760c24848ed6029b6b',
-  });
-
-  const ckbVirtualTxResult = await genTransferSporeCkbVirtualTx({
-    collector,
-    sporeRgbppLockArgs,
-    sporeTypeBytes,
-    isMainnet,
-  });
-
-  const { commitment, ckbRawTx } = ckbVirtualTxResult;
-
-  // console.log(JSON.stringify(ckbRawTx))
-
-  // Send BTC tx
-  const psbt = await sendRgbppUtxos({
-    ckbVirtualTx: ckbRawTx,
-    commitment,
-    tos: [toBtcAddress],
-    ckbCollector: collector,
-    from: btcAddress!,
-    source,
-    feeRate: 30,
-  });
-  psbt.signAllInputs(keyPair);
-  psbt.finalizeAllInputs();
-
-  const btcTx = psbt.extractTransaction();
-  const { txid: btcTxId } = await service.sendBtcTransaction(btcTx.toHex());
-
-  console.log('BTC TxId: ', btcTxId);
-
-  try {
-    await service.sendRgbppCkbTransaction({ btc_txid: btcTxId, ckb_virtual_result: ckbVirtualTxResult });
-    const interval = setInterval(async () => {
-      const { state, failedReason } = await service.getRgbppTransactionState(btcTxId);
-      console.log('state', state);
-      if (state === 'completed' || state === 'failed') {
-        clearInterval(interval);
-        if (state === 'completed') {
-          const { txhash: txHash } = await service.getRgbppTransactionHash(btcTxId);
-          console.info(`Rgbpp spore has been transferred on BTC and the related CKB tx hash is ${txHash}`);
-        } else {
-          console.warn(`Rgbpp CKB transaction failed and the reason is ${failedReason} `);
-        }
-      }
-    }, 30 * 1000);
-  } catch (error) {
-    console.error(error);
-  }
-};
-
-// Use your real BTC UTXO information on the BTC Testnet
-// rgbppLockArgs: outIndexU32 + btcTxId
-transferSpore({
-  // The spore rgbpp lock args is from 3-create-spore.ts
-  sporeRgbppLockArgs: buildRgbppLockArgs(2, 'd5868dbde4be5e49876b496449df10150c356843afb6f94b08f8d81f394bb350'),
-  toBtcAddress: 'tb1qhp9fh9qsfeyh0yhewgu27ndqhs5qlrqwau28m7',
-});
