+++ examples/rgbpp/tsconfig.json
@@ -21,6 +21,6 @@
     "skipLibCheck": true,
     "strict": true
   },
-  "include": ["local/**/*.ts", "queue/**/*.ts"],
+  "include": ["queue/**/*.ts", "xudt/**/*.ts", "core/*.ts"],
   "exclude": ["node_modules"]
 }
